+++ Assets/Scripts/GameLogic/TerrainType.cs
@@ -4,7 +4,10 @@ namespace Ventura.GameLogic
     public record TerrainType: GameLogicObject
     {
         private string _name;
-        public string Name { get { return _name; }  }
+        public string Name { get { return _name; } }
+
+        private string _label;
+        public string Label { get { return _label; } }
 
         private bool _walkable;
         public bool Walkable { get { return _walkable; } }
@@ -15,18 +18,19 @@ namespace Ventura.GameLogic
         //darkTile: TerrainAspect
         //lightTile: TerrainAspect
 
-        public TerrainType(string name, bool walkable, bool transparent)
+        public TerrainType(string name, string label, bool walkable, bool transparent)
         {
             this._name = name;
+            this._label = label;
             this._walkable = walkable;
             this._transparent = transparent;
         }
 
-        public static TerrainType Plains1 = new TerrainType("plains1", true, true);
-        public static TerrainType Plains2 = new TerrainType("plains2", true, true);
-        public static TerrainType Hills1 = new TerrainType("hills1", true, true);
-        public static TerrainType Hills2 = new TerrainType("hills2", true, true);
-        public static TerrainType Mountains = new TerrainType("mountains", true, true);
+        public static TerrainType Plains1 = new TerrainType("plains1", "Plains", true, true);
+        public static TerrainType Plains2 = new TerrainType("plains2", "Plains", true, true);
+        public static TerrainType Hills1 = new TerrainType("hills1", "Hills", true, true);
+        public static TerrainType Hills2 = new TerrainType("hills2", "Hills", true, true);
+        public static TerrainType Mountains = new TerrainType("mountains", "Mountains", true, true);
 
     }
 }